+++ plog_frontend/src/views/note/NoteDetail.vue
@@ -206,7 +206,12 @@ export default {
         getFormatDate(regtime) {
             return moment(new Date(regtime)).format('YYYY.MM.DD');
         },
-        deleteNote() {
+        deleteNote: async function () {
+          const res = await this.$dialog.warning({
+            text:"노트를 삭제 하시겠습니까?",
+            title: 'Delete Note'
+          });
+          if(res){
             http.delete('/post/', {
                 params : {
                   pId : this.pId,
@@ -214,7 +219,7 @@ export default {
               })
               .then((response) => {
                 if(response.data === 'success'){
-                  this.$dialog.notify.error("노트 삭제 완료 🤗", {
+                  this.$dialog.notify.success("노트 삭제 완료 🤗", {
                     position: "bottom-right",
                     timeout: 3000,
                   });
@@ -240,6 +245,7 @@ export default {
                     this.$router.push('/note')
                 }
               });
+          }
         },
         makePDF () {
             var fileName = this.Note.pTitle;
